refactor: split admin_telegram.py (3360 LOC → 6 modules)
Split god file backend/app/api/v1/endpoints/admin_telegram.py
(3360 LOC) into a package with 6 modules:

- __init__.py (55 LOC) — re-exports all names for backward compat
- _helpers.py (1761 LOC) — imports, constants, staff link token
  helpers, AI approval helpers, staff action validation
- _ai_approval.py (203 LOC) — AI approval alert + outcome endpoints
- _staff_actions.py (541 LOC) — staff action confirm endpoint,
  bot token status helpers, webhook info, command registration
- _settings.py (608 LOC) — settings CRUD, test-bot, register-commands,
  set-webhook, webhook-info, integration-status
- _management.py (247 LOC) — templates, test-message, stats, users,
  broadcast

Backward compatibility: __init__.py uses `from _helpers import *` +
explicit private name exports + imports from all submodules. External
imports like `from app.api.v1.endpoints.admin_telegram import
PATIENT_BOOKING_ENTRY_ROUTE` continue to work unchanged.

Tests: app loads OK; frontend 554/554 pass (0 regressions).
