+++ backend/app/api/v1/endpoints/admin_telegram/_helpers.py
@@ -1,3 +1,9 @@
+"""Admin Telegram helpers, constants, and utility functions.
+
+Split from admin_telegram.py (3343 LOC → modular).
+"""
+from __future__ import annotations
+
 """
 API endpoints для управления Telegram в админ панели
 """
@@ -1758,1602 +1764,3 @@ def _validate_staff_action_target_binding(
 # ============================================================
 # === AI APPROVAL ENDPOINTS ===
 # ============================================================
-
-@router.post("/telegram/ai-approval-alerts")
-async def send_telegram_ai_approval_alert(
-    request: TelegramAiApprovalAlertRequest,
-    db: Session = Depends(get_db),
-    current_user: User = Depends(
-        require_roles(
-            "Admin",
-            "Registrar",
-            "Cashier",
-            "Owner",
-            "Doctor",
-            "admin",
-            "registrar",
-            "cashier",
-            "owner",
-            "doctor",
-        )
-    ),
-) -> dict[str, Any]:
-    """Send one safe AI approval alert to a linked staff Telegram chat."""
-
-    workflow_key = str(request.workflow_key or "").strip().lower()
-    workflow = _telegram_ai_approval_workflow(workflow_key)
-    recipient_user = db.query(User).filter(User.id == request.recipient_user_id).first()
-    if not recipient_user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="ai_approval_recipient_not_found",
-        )
-    _assert_ai_approval_role_allowed(
-        workflow,
-        getattr(recipient_user, "role", None),
-        role_field="recipient_roles",
-    )
-
-    telegram_user = crud_telegram.get_telegram_user_by_linked_user_id(
-        db, request.recipient_user_id
-    )
-    if (
-        not telegram_user
-        or not bool(getattr(telegram_user, "active", False))
-        or bool(getattr(telegram_user, "blocked", False))
-    ):
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="linked_staff_telegram_chat_not_found",
-        )
-
-    protected_url = _protected_frontend_url(str(workflow["protected_route"]))
-    if not protected_url:
-        raise HTTPException(
-            status_code=status.HTTP_503_SERVICE_UNAVAILABLE,
-            detail="protected_frontend_url_not_configured",
-        )
-
-    message = build_telegram_ai_approval_message(
-        workflow_key,
-        protected_url,
-        request.metrics,
-    )
-    bot_service = await get_telegram_bot_service()
-    if not bool(getattr(bot_service, "active", False)):
-        await bot_service.initialize(db)
-    if not getattr(bot_service, "bot_token", None):
-        raise HTTPException(
-            status_code=status.HTTP_503_SERVICE_UNAVAILABLE,
-            detail="telegram_bot_token_not_configured",
-        )
-
-    success = bool(
-        await bot_service._send_message(
-            int(telegram_user.chat_id),
-            message["text"],
-            message["reply_markup"],
-        )
-    )
-    target_reference_hash = (
-        _telegram_ai_approval_reference_hash(request.target_reference)
-        if request.target_reference
-        else None
-    )
-    crud_audit.log(
-        db,
-        action=(
-            "telegram_ai_approval_notification_sent"
-            if success
-            else "telegram_ai_approval_notification_failed"
-        ),
-        entity_type="telegram_ai_approval",
-        entity_id=None,
-        actor_user_id=current_user.id,
-        payload={
-            "workflow_key": workflow_key,
-            "notification_type": workflow["notification_type"],
-            "recipient_user_id": request.recipient_user_id,
-            "recipient_role": _normalize_staff_role(getattr(recipient_user, "role", None)),
-            "target_reference_hash": target_reference_hash,
-            "safe_metric_keys": message["safe_metric_keys"],
-            "telegram_user_id_hash": _staff_runtime_reference_hash(
-                "telegram_chat", telegram_user.chat_id
-            ),
-            "plain_chat_medical_content_allowed": False,
-            "contains_plain_chat_medical_content": False,
-            "autonomous_mutation_allowed": False,
-            "domain_mutation": False,
-            "requires_human_confirmation": True,
-            "protected_route": workflow["protected_route"],
-            "timestamp": datetime.now(UTC).isoformat(),
-        },
-    )
-    db.commit()
-    if not success:
-        raise HTTPException(
-            status_code=status.HTTP_502_BAD_GATEWAY,
-            detail="telegram_ai_approval_notification_send_failed",
-        )
-
-    return {
-        "success": True,
-        "workflow_key": workflow_key,
-        "recipient_user_id": request.recipient_user_id,
-        "protected_route": workflow["protected_route"],
-        "safe_metric_keys": message["safe_metric_keys"],
-        "plain_chat_medical_content_allowed": False,
-        "autonomous_mutation_allowed": False,
-    }
-
-
-@router.post("/telegram/ai-approval-outcomes")
-def capture_telegram_ai_approval_outcome(
-    request: TelegramAiApprovalOutcomeRequest,
-    db: Session = Depends(get_db),
-    current_user: User = Depends(
-        require_roles(
-            "Admin",
-            "Registrar",
-            "Cashier",
-            "Owner",
-            "Doctor",
-            "admin",
-            "registrar",
-            "cashier",
-            "owner",
-            "doctor",
-        )
-    ),
-) -> dict[str, Any]:
-    """Capture human accepted/rejected feedback without mutating domain state."""
-
-    workflow_key = str(request.workflow_key or "").strip().lower()
-    workflow = _telegram_ai_approval_workflow(workflow_key)
-    _assert_ai_approval_role_allowed(
-        workflow,
-        getattr(current_user, "role", None),
-        role_field="outcome_roles",
-    )
-    outcome = _normalize_ai_approval_outcome(request.outcome)
-    reason_code = _normalize_ai_approval_reason_code(request.reason_code)
-    target_reference_hash = (
-        _telegram_ai_approval_reference_hash(request.target_reference)
-        if request.target_reference
-        else None
-    )
-
-    crud_audit.log(
-        db,
-        action="telegram_ai_approval_outcome_recorded",
-        entity_type="telegram_ai_approval",
-        entity_id=None,
-        actor_user_id=current_user.id,
-        payload={
-            "workflow_key": workflow_key,
-            "notification_type": workflow["notification_type"],
-            "outcome": outcome,
-            "reason_code": reason_code,
-            "actor_role": _normalize_staff_role(getattr(current_user, "role", None)),
-            "target_reference_hash": target_reference_hash,
-            "plain_chat_medical_content_allowed": False,
-            "contains_plain_chat_medical_content": False,
-            "autonomous_mutation_allowed": False,
-            "domain_mutation": False,
-            "requires_human_confirmation": True,
-            "timestamp": datetime.now(UTC).isoformat(),
-        },
-    )
-    db.commit()
-
-    return {
-        "success": True,
-        "workflow_key": workflow_key,
-        "outcome": outcome,
-        "reason_code": reason_code,
-        "target_reference_hash": target_reference_hash,
-        "plain_chat_medical_content_allowed": False,
-        "autonomous_mutation_allowed": False,
-        "domain_mutation": False,
-    }
-
-
-# ============================================================
-# === STAFF ACTION ENDPOINTS ===
-# ============================================================
-
-@router.post("/telegram/staff-actions/{confirmation_id}/confirm")
-def confirm_staff_action(
-    confirmation_id: int,
-    db: Session = Depends(get_db),
-    current_user: User = Depends(
-        require_roles(
-            "Admin",
-            "Registrar",
-            "Cashier",
-            "Owner",
-            "Doctor",
-            "Lab",
-            "admin",
-            "registrar",
-            "cashier",
-            "owner",
-            "doctor",
-            "lab",
-        )
-    ),
-    request: StaffActionConfirmRequest | None = Body(default=None),
-) -> dict[str, Any]:
-    """Confirm one explicitly enabled Telegram staff action in the protected app."""
-
-    if not isinstance(request, StaffActionConfirmRequest):
-        request = StaffActionConfirmRequest()
-    record = (
-        db.query(TelegramStaffConfirmationToken)
-        .filter(TelegramStaffConfirmationToken.id == confirmation_id)
-        .first()
-    )
-    if not record or int(record.staff_user_id) != int(current_user.id):
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="Staff action confirmation not found",
-        )
-
-    operation_key = str(record.operation_key or "")
-    command_key = str(record.command_key or "")
-    operation = _staff_state_change_operation_by_key(operation_key)
-    if not operation:
-        _record_staff_action_execution_failure(
-            db,
-            record=record,
-            current_user=current_user,
-            operation_key=operation_key,
-            command_key=command_key,
-            reason="operation_not_registered",
-        )
-        db.commit()
-        raise HTTPException(
-            status_code=status.HTTP_409_CONFLICT,
-            detail="Staff action operation is not registered",
-        )
-
-    role = _normalize_staff_role(getattr(current_user, "role", None))
-    allowed_roles = {
-        _normalize_staff_role(role_key) for role_key in operation.get("roles", [])
-    }
-    if role not in allowed_roles:
-        _record_staff_action_execution_failure(
-            db,
-            record=record,
-            current_user=current_user,
-            operation_key=operation_key,
-            command_key=command_key,
-            reason="role_not_allowed",
-        )
-        db.commit()
-        raise HTTPException(
-            status_code=status.HTTP_403_FORBIDDEN,
-            detail="Current role cannot confirm this staff action",
-        )
-
-    enabled_commands = {
-        str(item).strip().lower()
-        for item in STAFF_BOT_ACTION_ENABLEMENT_CONTRACT.get("enabled_commands", [])
-    }
-    if command_key.lower() not in enabled_commands:
-        _record_staff_action_execution_failure(
-            db,
-            record=record,
-            current_user=current_user,
-            operation_key=operation_key,
-            command_key=command_key,
-            reason="command_not_enabled",
-        )
-        db.commit()
-        raise HTTPException(
-            status_code=status.HTTP_409_CONFLICT,
-            detail="Staff action command is not enabled for protected execution",
-        )
-
-    target_error = _validate_staff_action_target_request(command_key, request)
-    if target_error:
-        _record_staff_action_execution_failure(
-            db,
-            record=record,
-            current_user=current_user,
-            operation_key=operation_key,
-            command_key=command_key,
-            reason=target_error,
-        )
-        db.commit()
-        raise HTTPException(
-            status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
-            detail=target_error,
-        )
-
-    target_binding_error = _validate_staff_action_target_binding(
-        record,
-        command_key,
-        request,
-    )
-    if target_binding_error:
-        _record_staff_action_execution_failure(
-            db,
-            record=record,
-            current_user=current_user,
-            operation_key=operation_key,
-            command_key=command_key,
-            reason=target_binding_error,
-        )
-        db.commit()
-        raise HTTPException(
-            status_code=status.HTTP_409_CONFLICT,
-            detail=target_binding_error,
-        )
-
-    consume_result = TelegramStaffConfirmationTokenService(db).consume_for_confirmation(
-        token_hash=record.token_hash,
-        staff_user_id=int(current_user.id),
-        telegram_chat_id=int(record.telegram_chat_id),
-        operation_key=record.operation_key,
-        action_payload_hash=record.action_payload_hash,
-        idempotency_key_hash=record.idempotency_key_hash,
-    )
-    if not consume_result.get("valid"):
-        _record_staff_action_execution_failure(
-            db,
-            record=record,
-            current_user=current_user,
-            operation_key=operation_key,
-            command_key=command_key,
-            reason=str(consume_result.get("reason") or "confirmation_rejected"),
-        )
-        db.commit()
-        raise HTTPException(
-            status_code=status.HTTP_409_CONFLICT,
-            detail={
-                "reason": consume_result.get("reason") or "confirmation_rejected",
-                "single_use_enforced": bool(
-                    consume_result.get("single_use_enforced")
-                ),
-            },
-        )
-
-    if command_key.lower() == "/call":
-        action_result = TelegramStaffActionAdapterService(db).staff_call_next_patient(
-            actor_user_id=int(current_user.id),
-            telegram_chat_id=int(record.telegram_chat_id),
-            commit=True,
-        )
-        return {
-            "success": True,
-            "confirmation_id": confirmation_id,
-            "operation_key": operation_key,
-            "command_key": command_key,
-            "action": action_result.get("action"),
-            "status": action_result.get("status"),
-            "queue_time_preserved": action_result.get("queue_time_preserved"),
-        }
-
-    if command_key.lower() == "/skip":
-        action_result = TelegramStaffActionAdapterService(db).staff_skip_queue_entry(
-            entry_id=int(request.entry_id),
-            actor_user_id=int(current_user.id),
-            telegram_chat_id=int(record.telegram_chat_id),
-            commit=True,
-        )
-        return {
-            "success": True,
-            "confirmation_id": confirmation_id,
-            "operation_key": operation_key,
-            "command_key": command_key,
-            "action": action_result.get("action"),
-            "status": action_result.get("status"),
-            "queue_time_preserved": action_result.get("queue_time_preserved"),
-        }
-
-    if command_key.lower() == "/cancel_visit":
-        action_result = TelegramStaffActionAdapterService(db).staff_cancel_visit(
-            visit_id=int(request.visit_id),
-            actor_user_id=int(current_user.id),
-            telegram_chat_id=int(record.telegram_chat_id),
-            commit=True,
-        )
-        return {
-            "success": True,
-            "confirmation_id": confirmation_id,
-            "operation_key": operation_key,
-            "command_key": command_key,
-            "action": action_result.get("action"),
-            "visit_status": action_result.get("visit_status"),
-        }
-
-    if command_key.lower() == "/move_visit":
-        action_result = TelegramStaffActionAdapterService(db).staff_move_visit(
-            visit_id=int(request.visit_id),
-            new_visit_date=request.new_visit_date,
-            actor_user_id=int(current_user.id),
-            telegram_chat_id=int(record.telegram_chat_id),
-            commit=True,
-        )
-        return {
-            "success": True,
-            "confirmation_id": confirmation_id,
-            "operation_key": operation_key,
-            "command_key": command_key,
-            "action": action_result.get("action"),
-            "visit_date": action_result.get("visit_date").isoformat()
-            if action_result.get("visit_date")
-            else None,
-        }
-
-    if command_key.lower() == "/payment_status":
-        action_result = (
-            TelegramStaffActionAdapterService(db).staff_change_payment_status(
-                payment_id=int(request.payment_id),
-                new_status=str(request.new_status),
-                actor_user_id=int(current_user.id),
-                telegram_chat_id=int(record.telegram_chat_id),
-                commit=True,
-            )
-        )
-        return {
-            "success": True,
-            "confirmation_id": confirmation_id,
-            "operation_key": operation_key,
-            "command_key": command_key,
-            "action": action_result.get("action"),
-            "status": action_result.get("status"),
-        }
-
-    if command_key.lower() == "/refund":
-        action_result = TelegramStaffActionAdapterService(db).staff_refund_payment(
-            payment_id=int(request.payment_id),
-            amount=request.refund_amount,
-            reason=request.refund_reason,
-            actor_user_id=int(current_user.id),
-            telegram_chat_id=int(record.telegram_chat_id),
-            commit=True,
-        )
-        return {
-            "success": True,
-            "confirmation_id": confirmation_id,
-            "operation_key": operation_key,
-            "command_key": command_key,
-            "action": action_result.get("action"),
-            "status": action_result.get("status"),
-        }
-
-    if command_key.lower() == "/change_schedule":
-        action_result = (
-            TelegramStaffActionAdapterService(db).staff_change_doctor_schedule(
-                schedule_id=int(request.schedule_id),
-                start_time=request.start_time,
-                end_time=request.end_time,
-                breaks=request.breaks,
-                active=request.active,
-                actor_user_id=int(current_user.id),
-                telegram_chat_id=int(record.telegram_chat_id),
-                commit=True,
-            )
-        )
-        return {
-            "success": True,
-            "confirmation_id": confirmation_id,
-            "operation_key": operation_key,
-            "command_key": command_key,
-            "action": action_result.get("action"),
-            "active": action_result.get("active"),
-        }
-
-    _record_staff_action_execution_failure(
-        db,
-        record=record,
-        current_user=current_user,
-        operation_key=operation_key,
-        command_key=command_key,
-        reason="enabled_command_missing_executor",
-    )
-    db.commit()
-    raise HTTPException(
-        status_code=status.HTTP_409_CONFLICT,
-        detail="Enabled staff action has no executor",
-    )
-
-
-def raise_admin_telegram_error(
-    action: str,
-    public_detail: str,
-    exc: Exception,
-    status_code: int = status.HTTP_500_INTERNAL_SERVER_ERROR,
-) -> NoReturn:
-    logger.warning(
-        "Admin Telegram endpoint failed action=%s status_code=%s error_type=%s",
-        action,
-        status_code,
-        type(exc).__name__,
-    )
-    raise HTTPException(status_code=status_code, detail=public_detail)
-
-
-def webhook_info_error_response(
-    action: str, public_error: str, exc: Exception
-) -> dict[str, Any]:
-    logger.warning(
-        "Admin Telegram webhook info failed action=%s error_type=%s",
-        action,
-        type(exc).__name__,
-    )
-    return {"webhook_set": False, "error": public_error}
-
-
-def _sanitize_telegram_webhook_info(webhook_info: dict[str, Any]) -> dict[str, Any]:
-    pending_update_count = webhook_info.get("pending_update_count")
-    if isinstance(pending_update_count, bool) or not isinstance(
-        pending_update_count, int
-    ):
-        pending_update_count = 0
-
-    return {
-        "url": str(webhook_info.get("url") or ""),
-        "pending_update_count": pending_update_count,
-    }
-
-
-class TelegramWebhookRequest(BaseModel):
-    webhook_url: str | None = None
-
-
-def _has_secret_value(value: Any) -> bool:
-    return bool(str(value or "").strip())
-
-
-def _get_configured_bot_token(db: Session) -> str | None:
-    config = crud_telegram.get_telegram_config(db)
-    if config and config.bot_token:
-        return config.bot_token
-
-    bot_token_setting = crud_clinic.get_setting_by_key(db, "bot_token")
-    return getattr(bot_token_setting, "value", None) if bot_token_setting else None
-
-
-def _build_staff_bot_token_status(
-    *,
-    token_value: Any,
-    source: str,
-    source_key: str,
-    patient_bot_token: str | None,
-) -> dict[str, Any]:
-    token_text = str(token_value or "").strip()
-    patient_token_text = str(patient_bot_token or "").strip()
-    token_present = bool(token_text)
-    patient_bot_token_reused = bool(
-        token_present and patient_token_text and token_text == patient_token_text
-    )
-    ready = token_present and not patient_bot_token_reused
-    return {
-        "configured": token_present,
-        "ready": ready,
-        "source": source,
-        "source_key": source_key,
-        "separate_staff_bot_token_configured": ready,
-        "patient_bot_token_reused": patient_bot_token_reused,
-    }
-
-
-def _get_staff_bot_token_runtime_status(
-    db: Session, patient_bot_token: str | None = None
-) -> dict[str, Any]:
-    for env_key in STAFF_BOT_TOKEN_ENV_KEYS:
-        token_value = os.getenv(env_key)
-        if _has_secret_value(token_value):
-            return _build_staff_bot_token_status(
-                token_value=token_value,
-                source="environment",
-                source_key=env_key,
-                patient_bot_token=patient_bot_token,
-            )
-
-    for setting_key in STAFF_BOT_TOKEN_SETTING_KEYS:
-        setting = crud_clinic.get_setting_by_key(db, setting_key)
-        token_value = getattr(setting, "value", None)
-        if _has_secret_value(token_value):
-            return _build_staff_bot_token_status(
-                token_value=token_value,
-                source="clinic_settings",
-                source_key=setting_key,
-                patient_bot_token=patient_bot_token,
-            )
-
-    return {
-        "configured": False,
-        "ready": False,
-        "source": "not_configured",
-        "source_key": None,
-        "separate_staff_bot_token_configured": False,
-        "patient_bot_token_reused": False,
-    }
-
-
-def _get_configured_staff_bot_token(
-    db: Session, patient_bot_token: str | None = None
-) -> str | None:
-    patient_token_text = str(patient_bot_token or "").strip()
-    for env_key in STAFF_BOT_TOKEN_ENV_KEYS:
-        token_text = str(os.getenv(env_key) or "").strip()
-        if token_text and token_text != patient_token_text:
-            return token_text
-
-    for setting_key in STAFF_BOT_TOKEN_SETTING_KEYS:
-        setting = crud_clinic.get_setting_by_key(db, setting_key)
-        token_text = str(getattr(setting, "value", None) or "").strip()
-        if token_text and token_text != patient_token_text:
-            return token_text
-
-    return None
-
-
-def _build_staff_bot_token_contract(token_status: dict[str, Any]) -> dict[str, Any]:
-    configured = bool(token_status.get("configured"))
-    ready = bool(token_status.get("ready"))
-    blocked_by = []
-    if not configured:
-        blocked_by.append("dedicated_staff_bot_token")
-    elif not ready:
-        blocked_by.append("separate_staff_bot_token_configured")
-    return {
-        **STAFF_BOT_TOKEN_CONTRACT,
-        "enabled": ready,
-        "runtime_read_enabled": True,
-        "configured": configured,
-        "ready": ready,
-        "separate_staff_bot_token_configured": ready,
-        "token_returned_to_frontend": False,
-        "patient_bot_token_reused": bool(token_status.get("patient_bot_token_reused")),
-        "patient_bot_transport_unchanged": True,
-        "source": token_status.get("source") or "not_configured",
-        "source_key": token_status.get("source_key"),
-        "accepted_env_keys": list(STAFF_BOT_TOKEN_ENV_KEYS),
-        "accepted_setting_keys": list(STAFF_BOT_TOKEN_SETTING_KEYS),
-        "runtime_blocked_by": blocked_by,
-        "checks": [
-            {
-                "key": "dedicated_staff_bot_token",
-                "ready": configured,
-            },
-            {
-                "key": "separate_staff_bot_token_configured",
-                "ready": ready,
-            },
-            {
-                "key": "token_not_returned_to_frontend",
-                "ready": True,
-            },
-            {
-                "key": "patient_bot_transport_unchanged",
-                "ready": True,
-            },
-        ],
-    }
-
-
-def _staff_bot_read_only_command_payload() -> list[dict[str, str]]:
-    commands = []
-    for item in STAFF_BOT_COMMAND_REGISTRATION_CONTRACT.get("commands", []):
-        if item.get("intent") != "read_only":
-            continue
-        command = str(item.get("command") or "").strip().lstrip("/")
-        description = str(item.get("label") or command).strip()
-        if command:
-            commands.append(
-                {
-                    "command": command,
-                    "description": description[:256],
-                }
-            )
-    return commands
-
-
-def _build_staff_command_registration_contract(
-    token_contract: dict[str, Any],
-) -> dict[str, Any]:
-    token_ready = bool(token_contract.get("ready"))
-    return {
-        **STAFF_BOT_COMMAND_REGISTRATION_CONTRACT,
-        "registration_enabled": token_ready,
-        "runtime_registration_enabled": True,
-        "registration_endpoint_published": True,
-        "read_only_command_registration_available": token_ready,
-        "read_only_commands_registered": False,
-        "registered_command_scope": "read_only_staff_commands_only",
-        "patient_bot_token_allowed": False,
-        "state_changing_commands_registered": False,
-        "runtime_blocked_by": (
-            [] if token_ready else list(token_contract["runtime_blocked_by"])
-        ),
-        "telegram_payload_commands": _staff_bot_read_only_command_payload(),
-    }
-
-
-def _get_configured_bot_username(db: Session) -> str | None:
-    config = crud_telegram.get_telegram_config(db)
-    if config and config.bot_username:
-        return config.bot_username
-
-    bot_username_setting = crud_clinic.get_setting_by_key(db, "bot_username")
-    return (
-        getattr(bot_username_setting, "value", None) if bot_username_setting else None
-    )
-
-
-def _fetch_telegram_webhook_info(bot_token: str) -> dict[str, Any]:
-    response = httpx.get(
-        f"https://api.telegram.org/bot{bot_token}/getWebhookInfo", timeout=10
-    )
-    response.raise_for_status()
-    result = response.json()
-    if not result.get("ok"):
-        raise RuntimeError(result.get("description") or "Telegram API error")
-    return result["result"]
-
-# ===================== НАСТРОЙКИ TELEGRAM =====================
-
-
-# ============================================================
-# === SETTINGS & BOT MANAGEMENT ENDPOINTS ===
-# ============================================================
-
-@router.get("/telegram/settings")
-def get_telegram_settings(
-    db: Session = Depends(get_db), current_user: User = Depends(require_roles("Admin"))
-):
-    """Получить настройки Telegram"""
-    try:
-        telegram_settings = crud_clinic.get_settings_by_category(db, "telegram")
-
-        result = {
-            "bot_token": "",
-            "webhook_url": "",
-            "admin_chat_ids": [],
-            "notifications_enabled": True,
-            "appointment_reminders": True,
-            "lab_results_notifications": True,
-            "payment_notifications": True,
-            "default_language": "ru",
-            "supported_languages": ["ru", "uz-Latn"],
-        }
-
-        # Применяем сохраненные настройки
-        for setting in telegram_settings:
-            if setting.key in result:
-                result[setting.key] = setting.value
-
-        # Скрываем токен бота в ответе
-        if result["bot_token"]:
-            result["bot_token_masked"] = "***скрыт***"
-            result["bot_token"] = "***скрыт***"
-
-        return result
-    except Exception as e:
-        raise_admin_telegram_error(
-            "settings-read",
-            "Ошибка получения настроек Telegram",
-            e,
-        )
-
-
-@router.put("/telegram/settings")
-def update_telegram_settings(
-    settings: dict[str, Any],
-    db: Session = Depends(get_db),
-    current_user: User = Depends(require_roles("Admin")),
-):
-    """Обновить настройки Telegram"""
-    try:
-        # TG-AUDIT-28 P0-5: фильтруем masked bot_token placeholder.
-        # GET /telegram/settings возвращает bot_token как "***скрыт***";
-        # фронтенд отправляет его обратно при Save; без фильтра backend
-        # записал бы masked значение в DB → бот ломается.
-        if settings and isinstance(settings, dict):
-            bt = settings.get("bot_token")
-            if bt and isinstance(bt, str) and "***" in bt:
-                settings = {k: v for k, v in settings.items() if k != "bot_token"}
-
-        # Обновляем настройки в категории "telegram"
-        updated_settings = crud_clinic.update_settings_batch(
-            db, "telegram", settings, current_user.id
-        )
-
-        return {
-            "success": True,
-            "message": "Настройки Telegram обновлены",
-            "updated_count": len(updated_settings),
-        }
-    except Exception as e:
-        raise_admin_telegram_error(
-            "settings-update",
-            "Ошибка обновления настроек Telegram",
-            e,
-        )
-
-
-@router.post("/telegram/test-bot")
-def test_telegram_bot(
-    db: Session = Depends(get_db), current_user: User = Depends(require_roles("Admin"))
-):
-    """Тестировать подключение к Telegram боту"""
-    try:
-        # Получаем токен бота
-        bot_token = _get_configured_bot_token(db)
-        if not bot_token:
-            raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST, detail="Токен бота не настроен"
-            )
-
-        # Тестируем подключение к API Telegram
-        response = httpx.get(
-            f"https://api.telegram.org/bot{bot_token}/getMe", timeout=10
-        )
-
-        if response.status_code == 200:
-            bot_info = response.json()
-            if bot_info.get("ok"):
-                bot_data = bot_info["result"]
-
-                # Сохраняем информацию о боте
-                crud_clinic.update_settings_batch(
-                    db,
-                    "telegram",
-                    {
-                        "bot_username": bot_data.get("username"),
-                        "bot_name": bot_data.get("first_name"),
-                    },
-                    current_user.id,
-                )
-                config_payload = {
-                    "bot_token": bot_token,
-                    "bot_username": bot_data.get("username"),
-                    "bot_name": bot_data.get("first_name"),
-                    "active": True,
-                }
-                if crud_telegram.get_telegram_config(db):
-                    crud_telegram.update_telegram_config(db, config_payload)
-                else:
-                    crud_telegram.create_telegram_config(db, config_payload)
-
-                return {
-                    "success": True,
-                    "message": "Подключение к боту успешно",
-                    "bot_info": {
-                        "id": bot_data.get("id"),
-                        "username": bot_data.get("username"),
-                        "first_name": bot_data.get("first_name"),
-                        "can_join_groups": bot_data.get("can_join_groups"),
-                        "can_read_all_group_messages": bot_data.get(
-                            "can_read_all_group_messages"
-                        ),
-                    },
-                }
-            else:
-                raise Exception(f"Ошибка API Telegram: {bot_info.get('description')}")
-        else:
-            raise Exception(f"HTTP {response.status_code}: {response.text}")
-
-    except requests.RequestException as e:
-        raise_admin_telegram_error(
-            "test-bot-request",
-            "Ошибка подключения к Telegram API",
-            e,
-            status.HTTP_400_BAD_REQUEST,
-        )
-    except HTTPException:
-        raise
-    except Exception as e:
-        raise_admin_telegram_error(
-            "test-bot",
-            "Ошибка тестирования бота",
-            e,
-        )
-
-
-@router.post("/telegram/register-patient-commands")
-async def register_patient_bot_commands(
-    db: Session = Depends(get_db),
-    current_user: User = Depends(require_roles("Admin")),
-):
-    """Register patient bot commands in Telegram via the configured bot token."""
-    try:
-        bot_token = _get_configured_bot_token(db)
-        if not bot_token:
-            raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                detail="Bot token is not configured",
-            )
-
-        bot_service = await get_telegram_bot_service()
-        bot_service.bot_token = bot_token
-        ok, error = await bot_service.set_patient_bot_commands()
-        if not ok:
-            raise HTTPException(
-                status_code=status.HTTP_502_BAD_GATEWAY,
-                detail={
-                    "message": "Telegram patient bot commands were not registered",
-                    "error": error,
-                },
-            )
-
-        return {
-            "success": True,
-            "message": (
-                "Telegram patient bot commands, menu button, and profile texts "
-                "registered"
-            ),
-            "registered_languages": ["ru", "uz"],
-            "menu_button": PATIENT_BOT_MENU_BUTTON,
-            "profile_texts": PATIENT_BOT_PROFILE_TEXTS,
-            "commands": {
-                "ru": PATIENT_BOT_COMMANDS_RU,
-                "uz": PATIENT_BOT_COMMANDS_UZ,
-            },
-        }
-    except HTTPException:
-        raise
-    except Exception as e:
-        raise_admin_telegram_error(
-            "register-patient-commands",
-            "Telegram patient bot command registration failed",
-            e,
-        )
-
-
-@router.post("/telegram/register-staff-commands")
-async def register_staff_bot_commands(
-    db: Session = Depends(get_db),
-    current_user: User = Depends(require_roles("Admin")),
-):
-    """Register read-only staff bot commands with a dedicated staff bot token."""
-    try:
-        patient_bot_token = _get_configured_bot_token(db)
-        token_status = _get_staff_bot_token_runtime_status(
-            db, patient_bot_token=patient_bot_token
-        )
-        staff_bot_token = _get_configured_staff_bot_token(
-            db, patient_bot_token=patient_bot_token
-        )
-        if not staff_bot_token or not token_status.get("ready"):
-            raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                detail={
-                    "message": "Dedicated staff bot token is not configured",
-                    "source": token_status.get("source"),
-                    "source_key": token_status.get("source_key"),
-                    "patient_bot_token_reused": bool(
-                        token_status.get("patient_bot_token_reused")
-                    ),
-                },
-            )
-
-        commands = _staff_bot_read_only_command_payload()
-        bot_service = await get_telegram_bot_service()
-        ok, error = await bot_service.set_staff_bot_commands(
-            staff_bot_token, commands=commands
-        )
-        if not ok:
-            raise HTTPException(
-                status_code=status.HTTP_502_BAD_GATEWAY,
-                detail={
-                    "message": "Telegram staff bot commands were not registered",
-                    "error": error,
-                },
-            )
-
-        return {
-            "success": True,
-            "message": "Telegram staff bot read-only commands registered",
-            "registered_commands": [item["command"] for item in commands],
-            "state_changing_commands_registered": False,
-            "token_returned_to_frontend": False,
-        }
-    except HTTPException:
-        raise
-    except Exception as e:
-        raise_admin_telegram_error(
-            "register-staff-commands",
-            "Telegram staff bot command registration failed",
-            e,
-        )
-
-
-@router.post("/telegram/set-webhook")
-def set_telegram_webhook(
-    payload: TelegramWebhookRequest | None = Body(default=None),
-    webhook_url: str | None = Query(default=None),
-    db: Session = Depends(get_db),
-    current_user: User = Depends(require_roles("Admin")),
-):
-    """Установить webhook для Telegram бота"""
-    try:
-        selected_webhook_url = (payload.webhook_url if payload else None) or webhook_url
-        if not selected_webhook_url:
-            raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                detail="webhook_url is required",
-            )
-
-        # Получаем токен бота
-        bot_token = _get_configured_bot_token(db)
-        if not bot_token:
-            raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST, detail="Токен бота не настроен"
-            )
-
-        secret_token = secrets.token_urlsafe(32)
-
-        # Устанавливаем webhook
-        response = httpx.post(
-            f"https://api.telegram.org/bot{bot_token}/setWebhook",
-            json={"url": selected_webhook_url, "secret_token": secret_token},
-            timeout=10,
-        )
-
-        if response.status_code == 200:
-            result = response.json()
-            if result.get("ok"):
-                # Сохраняем URL webhook
-                crud_clinic.update_setting(
-                    db, "webhook_url", {"value": selected_webhook_url}, current_user.id
-                )
-                config_payload = {
-                    "bot_token": bot_token,
-                    "bot_username": _get_configured_bot_username(db),
-                    "webhook_url": selected_webhook_url,
-                    "webhook_secret": secret_token,
-                    "active": True,
-                }
-                if crud_telegram.get_telegram_config(db):
-                    crud_telegram.update_telegram_config(db, config_payload)
-                else:
-                    crud_telegram.create_telegram_config(db, config_payload)
-
-                return {
-                    "success": True,
-                    "message": "Webhook установлен успешно",
-                    "webhook_url": selected_webhook_url,
-                    "webhook_secret_configured": True,
-                }
-            else:
-                raise Exception(
-                    f"Ошибка установки webhook: {result.get('description')}"
-                )
-        else:
-            raise Exception(f"HTTP {response.status_code}: {response.text}")
-
-    except requests.RequestException as e:
-        raise_admin_telegram_error(
-            "set-webhook-request",
-            "Ошибка установки webhook",
-            e,
-            status.HTTP_400_BAD_REQUEST,
-        )
-    except HTTPException:
-        raise
-    except Exception as e:
-        raise_admin_telegram_error(
-            "set-webhook",
-            "Ошибка установки webhook",
-            e,
-        )
-
-
-@router.get("/telegram/webhook-info")
-def get_telegram_webhook_info(
-    db: Session = Depends(get_db), current_user: User = Depends(require_roles("Admin"))
-):
-    """Получить информацию о webhook"""
-    try:
-        # Получаем токен бота
-        bot_token = _get_configured_bot_token(db)
-        if not bot_token:
-            return {"webhook_set": False, "message": "Токен бота не настроен"}
-
-        # Получаем информацию о webhook
-        response = httpx.get(
-            f"https://api.telegram.org/bot{bot_token}/getWebhookInfo", timeout=10
-        )
-
-        if response.status_code == 200:
-            result = response.json()
-            if result.get("ok"):
-                webhook_info = result["result"]
-                return {
-                    "webhook_set": bool(webhook_info.get("url")),
-                    "webhook_info": _sanitize_telegram_webhook_info(webhook_info),
-                }
-            else:
-                raise Exception(f"Ошибка API: {result.get('description')}")
-        else:
-            raise Exception(f"HTTP {response.status_code}")
-
-    except requests.RequestException as e:
-        return webhook_info_error_response(
-            "webhook-info-request",
-            "Ошибка подключения",
-            e,
-        )
-    except Exception as e:
-        return webhook_info_error_response(
-            "webhook-info",
-            "Ошибка получения информации о webhook",
-            e,
-        )
-
-
-# ===================== ШАБЛОНЫ СООБЩЕНИЙ =====================
-
-
-@router.get("/telegram/integration-status")
-def get_telegram_integration_status(
-    db: Session = Depends(get_db), current_user: User = Depends(require_roles("Admin"))
-):
-    """Return app-facing Telegram integration status without exposing secrets."""
-    try:
-        config = crud_telegram.get_telegram_config(db)
-        bot_token = _get_configured_bot_token(db)
-        bot_username = _get_configured_bot_username(db)
-        staff_bot_token_status = _get_staff_bot_token_runtime_status(
-            db, patient_bot_token=bot_token
-        )
-        telegram_users = crud_telegram.get_telegram_users(
-            db, active_only=False, limit=100000
-        )
-        linked_users = [user for user in telegram_users if user.patient_id]
-
-        webhook_info = None
-        webhook_error = None
-        if bot_token:
-            try:
-                webhook_info = _fetch_telegram_webhook_info(bot_token)
-            except requests.RequestException as exc:
-                webhook_error = "Telegram API unavailable"
-                logger.warning(
-                    "Admin Telegram integration status request failed error_type=%s",
-                    type(exc).__name__,
-                )
-            except Exception as exc:
-                webhook_error = "Telegram webhook status unavailable"
-                logger.warning(
-                    "Admin Telegram integration status failed error_type=%s",
-                    type(exc).__name__,
-                )
-
-        webhook_set = bool(webhook_info and webhook_info.get("url"))
-        webhook_url = (
-            webhook_info.get("url")
-            if webhook_info
-            else getattr(config, "webhook_url", None)
-        )
-
-        return {
-            "configured": bool(bot_token),
-            "active": bool(getattr(config, "active", False) or bot_token),
-            "bot_username": bot_username,
-            "mode": "webhook" if webhook_set else "polling",
-            "polling_ready": bool(bot_token and not webhook_set),
-            "polling_command": "python -m app.scripts.telegram_polling_worker",
-            "polling_task_name": "KosmedTelegramPollingWorker",
-            "webhook_set": webhook_set,
-            "webhook_url": webhook_url,
-            "webhook_error": webhook_error,
-            "pending_update_count": (
-                webhook_info.get("pending_update_count") if webhook_info else None
-            ),
-            "qr_linking_enabled": bool(bot_username),
-            "contact_linking_enabled": bool(bot_token),
-            "linked_users": len(linked_users),
-            "total_users": len(telegram_users),
-            "supported_functions": [
-                "ticket_qr_link",
-                "contact_phone_link",
-                "patient_queue",
-                "patient_payments_debt",
-                "patient_payments_mini_app_entry",
-                "patient_status",
-                "patient_language_notification_settings",
-                "lab_results_pdf",
-                "staff_link_start_token_handler",
-                "telegram_ai_approval_notifications",
-                "telegram_ai_approval_outcome_capture",
-            ],
-            "planned_functions": [
-                "dedicated_staff_bot_token_readiness_contract",
-                "staff_read_only_menu_contract",
-                "staff_role_menu_enablement_status",
-                "staff_role_linking_contract",
-                "staff_role_linking_runtime",
-                "staff_link_token_validation_contract",
-                "staff_link_token_validation_runtime_helper",
-                "staff_link_token_storage_migration_contract",
-                "staff_server_side_authorization_contract",
-                "staff_command_registration_contract",
-                "staff_state_change_confirmation_contract",
-                "staff_audit_logging_contract",
-                "staff_role_menus",
-                "staff_action_confirmations",
-                "staff_audit_logging",
-                "admin_notifications",
-            ],
-            "ai_approval": _build_telegram_ai_approval_status(),
-            "patient_bot": {
-                "version": "v1",
-                "transport": "polling" if not webhook_set else "webhook",
-                "supported_languages": [
-                    {"code": "ru", "label": "Русский"},
-                    {"code": "uz-Latn", "label": "O'zbekcha"},
-                ],
-                "default_language": "ru",
-                "onboarding": "language_choice_then_contact_link",
-                "commands": [
-                    {
-                        "command": f"/{command['command']}",
-                        "label": command["description"],
-                    }
-                    for command in PATIENT_BOT_COMMANDS_RU
-                ],
-                "features": [
-                    {
-                        "key": "ticket_qr_link",
-                        "label": "Привязка через QR чека",
-                        "enabled": bool(bot_username),
-                    },
-                    {
-                        "key": "contact_phone_link",
-                        "label": "Привязка через номер телефона",
-                        "enabled": bool(bot_token),
-                    },
-                    {
-                        "key": "patient_booking_entrypoint",
-                        "label": "Безопасная подсказка для записи",
-                        "enabled": bool(bot_token),
-                        "contract": PATIENT_BOOKING_ENTRY_CONTRACT,
-                    },
-                    {
-                        "key": "patient_services_menu",
-                        "label": "Видимая карта функций бота",
-                        "enabled": bool(bot_token),
-                    },
-                    {
-                        "key": "patient_queue",
-                        "label": "Очередь пациента на сегодня",
-                        "enabled": bool(bot_token),
-                    },
-                    {
-                        "key": "patient_visits",
-                        "label": "Мои визиты без медицинских деталей",
-                        "enabled": bool(bot_token),
-                    },
-                    {
-                        "key": "patient_payments_debt",
-                        "label": "Оплаты и долг по визиту",
-                        "enabled": bool(bot_token),
-                    },
-                    {
-                        "key": "patient_payments_mini_app_entry",
-                        "label": "Защищенный вход к оплатам пациента",
-                        "enabled": bool(bot_token),
-                        "contract": PATIENT_PAYMENT_ENTRY_CONTRACT,
-                    },
-                    {
-                        "key": "patient_mini_app_manifest",
-                        "label": "Mini App manifest пациента",
-                        "enabled": bool(bot_token),
-                        "contract": PATIENT_MINI_APP_MANIFEST_CONTRACT,
-                    },
-                    {
-                        "key": "lab_results_pdf",
-                        "label": "PDF-результаты лаборатории",
-                        "enabled": bool(bot_token),
-                    },
-                    {
-                        "key": "patient_forms_placeholder",
-                        "label": "Анкеты пациента: безопасная заглушка Mini App",
-                        "enabled": bool(bot_token),
-                        "contract": PATIENT_FORMS_ENTRY_CONTRACT,
-                    },
-                    {
-                        "key": "patient_documents_placeholder",
-                        "label": "Документы и чеки: будущий защищенный кабинет",
-                        "enabled": bool(bot_token),
-                    },
-                    {
-                        "key": "doctor_schedule_placeholder",
-                        "label": "Врачи и расписание: безопасная подсказка",
-                        "enabled": bool(bot_token),
-                    },
-                    {
-                        "key": "patient_cabinet_placeholder",
-                        "label": "Кабинет пациента: будущий защищенный вход",
-                        "enabled": bool(bot_token),
-                    },
-                    {
-                        "key": "patient_language_notification_settings",
-                        "label": "Настройки языка и уведомлений",
-                        "enabled": bool(bot_token),
-                    },
-                    {
-                        "key": "patient_support_contact",
-                        "label": "Безопасная связь с клиникой",
-                        "enabled": bool(bot_token),
-                    },
-                    {
-                        "key": "staff_entry_placeholder",
-                        "label": "Режим сотрудника: вход по персональной ссылке",
-                        "enabled": bool(bot_token),
-                    },
-                ],
-                "results_delivery": "telegram_pdf",
-                "max_pdf_reports_per_request": 3,
-            },
-            "staff_bot": _build_staff_bot_status(webhook_set, staff_bot_token_status),
-            "transition_path": (
-                "Set webhook when a public HTTPS backend URL is available; "
-                "stop polling before webhook is enabled."
-            ),
-        }
-    except Exception as e:
-        raise_admin_telegram_error(
-            "integration-status",
-            "Ошибка получения статуса Telegram интеграции",
-            e,
-        )
-
-
-# ============================================================
-# === TEMPLATES & MESSAGING ENDPOINTS ===
-# ============================================================
-
-@router.get("/telegram/templates")
-def get_telegram_templates(
-    language: str = "ru",
-    template_type: str | None = None,
-    db: Session = Depends(get_db),
-    current_user: User = Depends(require_roles("Admin")),
-):
-    """Получить шаблоны сообщений"""
-    try:
-        # Здесь будет логика получения шаблонов из БД
-        # Пока возвращаем базовые шаблоны
-
-        base_templates = {
-            "appointment_reminder": {
-                "ru": {
-                    "subject": "Напоминание о приеме",
-                    "message_text": "Здравствуйте, {{patient_name}}!\n\nНапоминаем о приеме {{appointment_date}} в {{appointment_time}} у врача {{doctor_name}} ({{specialty}}).\n\nАдрес: {{clinic_address}}\nКабинет: {{cabinet}}\n\nПодтвердите или выберите действие:",
-                    "buttons": [
-                        {
-                            "text": "✅ Подтвердить",
-                            "callback_data": "confirm_{{appointment_id}}",
-                        },
-                        {
-                            "text": "🔁 Перенести",
-                            "callback_data": "reschedule_{{appointment_id}}",
-                        },
-                        {
-                            "text": "❌ Отменить",
-                            "callback_data": "cancel_{{appointment_id}}",
-                        },
-                    ],
-                },
-                "uz": {
-                    "subject": "Qabul haqida eslatma",
-                    "message_text": "Assalomu alaykum, {{patient_name}}!\n\n{{appointment_date}} kuni soat {{appointment_time}}da {{doctor_name}} shifokorining qabuliga eslatma ({{specialty}}).\n\nManzil: {{clinic_address}}\nXona: {{cabinet}}\n\nTasdiqlang yoki amalni tanlang:",
-                    "buttons": [
-                        {
-                            "text": "✅ Tasdiqlash",
-                            "callback_data": "confirm_{{appointment_id}}",
-                        },
-                        {
-                            "text": "🔁 Ko'chirish",
-                            "callback_data": "reschedule_{{appointment_id}}",
-                        },
-                        {
-                            "text": "❌ Bekor qilish",
-                            "callback_data": "cancel_{{appointment_id}}",
-                        },
-                    ],
-                },
-            },
-            "lab_results_ready": {
-                "ru": {
-                    "subject": "Результаты анализов готовы",
-                    "message_text": "Здравствуйте, {{patient_name}}!\n\nГотовы результаты анализов:\n{{lab_tests}}\n\nВы можете получить их в клинике или скачать по ссылке ниже.",
-                    "buttons": [
-                        {"text": "📄 Скачать результаты", "url": "{{download_link}}"}
-                    ],
-                },
-                "uz": {
-                    "subject": "Tahlil natijalari tayyor",
-                    "message_text": "Assalomu alaykum, {{patient_name}}!\n\nTahlil natijalari tayyor:\n{{lab_tests}}\n\nUlarni klinikadan olishingiz yoki quyidagi havoladan yuklab olishingiz mumkin.",
-                    "buttons": [
-                        {
-                            "text": "📄 Natijalarni yuklab olish",
-                            "url": "{{download_link}}",
-                        }
-                    ],
-                },
-            },
-            "payment_confirmation": {
-                "ru": {
-                    "subject": "Платеж подтвержден",
-                    "message_text": "Платеж на сумму {{amount}} {{currency}} успешно обработан.\n\nСпасибо за оплату!\n\nЧек можете скачать по ссылке ниже.",
-                    "buttons": [{"text": "🧾 Скачать чек", "url": "{{receipt_link}}"}],
-                }
-            },
-        }
-
-        if template_type:
-            return base_templates.get(template_type, {}).get(language, {})
-
-        # Возвращаем все шаблоны для языка
-        result = {}
-        for template_key, template_data in base_templates.items():
-            if language in template_data:
-                result[template_key] = template_data[language]
-
-        return result
-
-    except Exception as e:
-        raise_admin_telegram_error(
-            "templates",
-            "Ошибка получения шаблонов",
-            e,
-        )
-
-
-@router.post("/telegram/send-test-message")
-def send_test_message(
-    chat_id: int,
-    message: str,
-    db: Session = Depends(get_db),
-    current_user: User = Depends(require_roles("Admin")),
-):
-    """Отправить тестовое сообщение"""
-    try:
-        # Получаем токен бота
-        bot_token = _get_configured_bot_token(db)
-        if not bot_token:
-            raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST, detail="Токен бота не настроен"
-            )
-
-        # Отправляем сообщение
-        response = httpx.post(
-            f"https://api.telegram.org/bot{bot_token}/sendMessage",
-            json={
-                "chat_id": chat_id,
-                "text": f"🧪 Тестовое сообщение от Programma Clinic\n\n{message}\n\n⚙️ Отправлено из админ панели",
-                "parse_mode": "HTML",
-            },
-            timeout=10,
-        )
-
-        if response.status_code == 200:
-            result = response.json()
-            if result.get("ok"):
-                return {
-                    "success": True,
-                    "message": "Тестовое сообщение отправлено",
-                    "message_id": result["result"]["message_id"],
-                }
-            else:
-                raise Exception(f"Ошибка отправки: {result.get('description')}")
-        else:
-            raise Exception(f"HTTP {response.status_code}: {response.text}")
-
-    except requests.RequestException as e:
-        raise_admin_telegram_error(
-            "send-test-message-request",
-            "Ошибка отправки сообщения",
-            e,
-            status.HTTP_400_BAD_REQUEST,
-        )
-    except HTTPException:
-        raise
-    except Exception as e:
-        raise_admin_telegram_error(
-            "send-test-message",
-            "Ошибка отправки тестового сообщения",
-            e,
-        )
-
-
-# ===================== СТАТИСТИКА TELEGRAM =====================
-
-
-@router.get("/telegram/stats")
-def get_telegram_stats(
-    days_back: int = 7,
-    db: Session = Depends(get_db),
-    current_user: User = Depends(require_roles(["Admin", "SuperAdmin", "admin"])),
-):
-    """Получить статистику Telegram"""
-    try:
-        # Пока возвращаем заглушку, в будущем будет реальная статистика
-        return {
-            "total_users": 0,
-            "active_users": 0,
-            "messages_sent": 0,
-            "messages_delivered": 0,
-            "messages_failed": 0,
-            "by_message_type": {},
-            "period_start": datetime.now(UTC) - timedelta(days=days_back),
-            "period_end": datetime.now(UTC),
-        }
-    except Exception as e:
-        raise_admin_telegram_error(
-            "stats",
-            "Ошибка получения статистики Telegram",
-            e,
-        )
-
-
-# ===================== УПРАВЛЕНИЕ ПОЛЬЗОВАТЕЛЯМИ =====================
-
-
-@router.get("/telegram/users")
-def get_telegram_users(
-    skip: int = 0,
-    limit: int = 100,
-    active_only: bool = False,
-    db: Session = Depends(get_db),
-    current_user: User = Depends(require_roles(["Admin", "SuperAdmin", "admin"])),
-):
-    """Получить список пользователей Telegram"""
-    try:
-        # Пока возвращаем заглушку
-        return []
-    except Exception as e:
-        raise_admin_telegram_error(
-            "users",
-            "Ошибка получения пользователей Telegram",
-            e,
-        )
-
-
-# ===================== ШИРОКОВЕЩАТЕЛЬНЫЕ СООБЩЕНИЯ =====================
-
-
-@router.post("/telegram/broadcast")
-def send_broadcast_message(
-    message: str,
-    target_groups: list[str],  # ["patients", "doctors", "admins"]
-    language: str = "ru",
-    db: Session = Depends(get_db),
-    current_user: User = Depends(require_roles("Admin")),
-):
-    """Отправить широковещательное сообщение"""
-    try:
-        # Получаем токен бота
-        bot_token = _get_configured_bot_token(db)
-        if not bot_token:
-            raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST, detail="Токен бота не настроен"
-            )
-
-        # Здесь будет логика отправки широковещательных сообщений
-        # Пока возвращаем заглушку
-
-        return {
-            "success": True,
-            "message": "Широковещательное сообщение поставлено в очередь",
-            "target_groups": target_groups,
-            "estimated_recipients": 0,  # Будет подсчитано из БД
-        }
-
-    except HTTPException:
-        raise
-    except Exception as e:
-        raise_admin_telegram_error(
-            "broadcast",
-            "Ошибка отправки сообщения",
-            e,
-        )
